When expanding width, we move to the top right corner of the screen.
Expanding width and minimizing both cause us to store
our current position prior to the operation.

The different states and how they interact here is starting
to get complex.  What keeps things simple so far is that
the states are on/off toggles - and we don't have too many
of them.